chore: Refactor codemods (#8056)
* initial codemod refactor

* fix TableView

* fix selectionStyle

* fix key to id

* clean up comments

* lint

* clean up

* update README.md

* improve transform function/option names

* update action group comments
