+++ packages/dev/codemods/src/s1-to-s2/README.md
@@ -13,3 +13,11 @@ Run `npx @react-spectrum/codemods s1-to-s2` from the directory you want to upgra
 ## How it works
 
 The upgrade assistant use codemods written with [jscodeshift](https://github.com/facebook/jscodeshift).
+
+## Adding a new codemod
+
+To add a new codemod for `Button`, for example, you would:
+
+1. Create a new transform function in `src/codemods/components/Button/transform.ts` and export it as `default`
+2. Implement the transform logic
+3. Add tests for the transform in `__tests__/button.test.ts`
