+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/ActionMenu/transform.ts
@@ -1,0 +1,16 @@
+import {commentOutProp} from '../../shared/transforms';
+import {NodePath} from '@babel/traverse';
+import * as t from '@babel/types';
+
+/**
+ * Transforms ActionMenu:
+ * - Comment out closeOnSelect (it has not been implemented yet).
+ * - Comment out trigger (it has not been implemented yet).
+ */
+export default function transformActionMenu(path: NodePath<t.JSXElement>) {
+  // Comment out closeOnSelect
+  commentOutProp(path, {propName: 'closeOnSelect'});
+
+  // Comment out trigger
+  commentOutProp(path, {propName: 'trigger'});
+}
