+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/Badge/transform.ts
@@ -1,0 +1,17 @@
+import {NodePath} from '@babel/traverse';
+import * as t from '@babel/types';
+import {updatePropNameAndValue} from '../../shared/transforms';
+
+/**
+ * Transforms Badge:
+ * - Change variant="info" to variant="informative".
+ */
+export default function transformBadge(path: NodePath<t.JSXElement>) {
+  // Change variant="info" to variant="informative"
+  updatePropNameAndValue(path, {
+    oldPropName: 'variant',
+    oldPropValue: 'info',
+    newPropName: 'variant',
+    newPropValue: 'informative'
+  });
+}
