+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/Button/transform.ts
@@ -1,0 +1,61 @@
+import {
+  commentOutProp,
+  removeProp,
+  updateComponentIfPropPresent,
+  updatePropName,
+  updatePropNameAndValue,
+  updatePropValueAndAddNewPropName
+} from '../../shared/transforms';
+import {NodePath} from '@babel/traverse';
+import * as t from '@babel/types';
+
+/**
+ * Transforms Button:
+ * - Change variant="cta" to variant="accent"
+ * - Change variant="overBackground" to variant="primary" staticColor="white"
+ * - Change style to fillStyle
+ * - Comment out isPending (it has not been implemented yet)
+ * - Remove isQuiet (it is no longer supported in Spectrum 2)
+ * - If href is present, the Button should be converted to a LinkButton
+ * - Remove elementType (it is no longer supported in Spectrum 2).
+ */
+export default function transformButton(path: NodePath<t.JSXElement>) {
+  // Change variant="cta" to variant="accent"
+  updatePropNameAndValue(path, {
+    oldPropName: 'variant',
+    oldPropValue: 'cta',
+    newPropName: 'variant',
+    newPropValue: 'accent'
+  });
+
+  // Change variant="overBackground" to variant="primary" staticColor="white"
+  updatePropValueAndAddNewPropName(path, {
+    oldPropName: 'variant',
+    oldPropValue: 'overBackground',
+    newPropName: 'variant',
+    newPropValue: 'primary',
+    additionalPropName: 'staticColor',
+    additionalPropValue: 'white'
+  });
+
+  // Change style to fillStyle
+  updatePropName(path, {
+    oldPropName: 'style',
+    newPropName: 'fillStyle'
+  });
+
+  // Comment out isPending
+  commentOutProp(path, {propName: 'isPending'});
+
+  // Remove isQuiet
+  removeProp(path, {propName: 'isQuiet'});
+
+  // If href is present, the Button should be converted to a LinkButton
+  updateComponentIfPropPresent(path, {
+    propName: 'href',
+    newComponentName: 'LinkButton'
+  });
+
+  // Remove elementType
+  removeProp(path, {propName: 'elementType'});
+}
