+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/CheckboxGroup/transform.ts
@@ -1,0 +1,12 @@
+import {NodePath} from '@babel/traverse';
+import {removeProp} from '../../shared/transforms';
+import * as t from '@babel/types';
+
+/**
+ * Transforms CheckboxGroup:
+ * - Remove showErrorIcon (it has been removed due to accessibility issues).
+ */
+export default function transformCheckboxGroup(path: NodePath<t.JSXElement>) {
+  // Remove showErrorIcon
+  removeProp(path, {propName: 'showErrorIcon'});
+}
