+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/ColorSlider/transform.ts
@@ -1,0 +1,12 @@
+import {NodePath} from '@babel/traverse';
+import {removeProp} from '../../shared/transforms';
+import * as t from '@babel/types';
+
+/**
+ * Transforms ColorSlider:
+ * - Remove showValueLabel (it has been removed due to accessibility issues).
+ */
+export default function transformColorSlider(path: NodePath<t.JSXElement>) {
+  // Remove showValueLabel
+  removeProp(path, {propName: 'showValueLabel'});
+}
