+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/Column/transform.ts
@@ -1,0 +1,12 @@
+import {NodePath} from '@babel/traverse';
+import * as t from '@babel/types';
+import {updateKeyToId} from '../../shared/transforms';
+
+/**
+ * Transforms Column:
+ * - Update key to id.
+ */
+export default function transformColumn(path: NodePath<t.JSXElement>) {
+  // Update key to id
+  updateKeyToId(path);
+}
