+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/Divider/transform.ts
@@ -1,0 +1,12 @@
+import {NodePath} from '@babel/traverse';
+import {removeComponentIfWithinParent} from '../../shared/transforms';
+import * as t from '@babel/types';
+
+/**
+ * Transforms Divider:
+ * - Remove Divider component if within a Dialog (Updated design for Dialog in Spectrum 2).
+ */
+export default function transformDivider(path: NodePath<t.JSXElement>) {
+  // Remove Divider component if within a Dialog
+  removeComponentIfWithinParent(path, {parentComponentName: 'Dialog'});
+}
