+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/NumberField/transform.ts
@@ -1,0 +1,25 @@
+import {NodePath} from '@babel/traverse';
+import {removeProp, updatePropNameAndValue} from '../../shared/transforms';
+import * as t from '@babel/types';
+
+/**
+ * Transforms NumberField:
+ * - Remove isQuiet (it is no longer supported in Spectrum 2).
+ * - Change validationState="invalid" to isInvalid.
+ * - Remove validationState="valid" (it is no longer supported in Spectrum 2).
+ */
+export default function transformNumberField(path: NodePath<t.JSXElement>) {
+  // Remove isQuiet
+  removeProp(path, {propName: 'isQuiet'});
+
+  // Change validationState="invalid" to isInvalid
+  updatePropNameAndValue(path, {
+    oldPropName: 'validationState',
+    oldPropValue: 'invalid',
+    newPropName: 'isInvalid',
+    newPropValue: true
+  });
+
+  // Remove validationState="valid"
+  removeProp(path, {propName: 'validationState', propValue: 'valid'});
+}
