+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/RadioGroup/transform.ts
@@ -1,0 +1,25 @@
+import {NodePath} from '@babel/traverse';
+import {removeProp, updatePropNameAndValue} from '../../shared/transforms';
+import * as t from '@babel/types';
+
+/**
+ * Transforms RadioGroup:
+ * - Change validationState="invalid" to isInvalid.
+ * - Remove validationState="valid" (it is no longer supported in Spectrum 2).
+ * - Remove showErrorIcon (it has been removed due to accessibility issues).
+ */
+export default function transformRadioGroup(path: NodePath<t.JSXElement>) {
+  // Change validationState="invalid" to isInvalid
+  updatePropNameAndValue(path, {
+    oldPropName: 'validationState',
+    oldPropValue: 'invalid',
+    newPropName: 'isInvalid',
+    newPropValue: true
+  });
+
+  // Remove validationState="valid"
+  removeProp(path, {propName: 'validationState', propValue: 'valid'});
+
+  // Remove showErrorIcon
+  removeProp(path, {propName: 'showErrorIcon'});
+}
