+++ packages/dev/codemods/src/s1-to-s2/src/codemods/components/TooltipTrigger/transform.ts
@@ -1,0 +1,15 @@
+import {NodePath} from '@babel/traverse';
+import * as t from '@babel/types';
+import {updatePlacementToSingleValue} from '../../shared/transforms';
+
+/**
+ * Transforms TooltipTrigger:
+ * - Updates placement prop to single value.
+ */
+export default function transformTooltipTrigger(path: NodePath<t.JSXElement>) {
+  // Update placement prop to single value
+  updatePlacementToSingleValue(path, {
+    propToUpdateName: 'placement',
+    childComponentName: 'Tooltip'
+  });
+}
