+++ packages/dev/codemods/src/s1-to-s2/src/codemods/icons/iconMap.ts
